Update spinningSurface.py
now there's a light source, but light diffusion is only based on distance, and not on the normal vector of the surface combined with the direction of the light. Also, light diffusion is scaled up to the 5th power to make it noticeable, also the whole thing is hecka slow. I'd love to have a density of 100, but noooOOOOoooo.